pyproject.toml
[project]
name = "rest-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "pcm422",email = "pcm0422@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
